requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coverage==5.3
docopt==0.6.2
guizero==1.1.0
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pathtools==0.1.2
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-describe==1.0.0
pytest-testmon==1.0.3
pytest-watch==4.2.0
six==1.15.0
tk==0.1.0
toml==0.10.2
watchdog==0.10.3
zipp==3.4.0
